chore: update migrations after 3.9.0 release
