+++ crates/contract-history/src/lib.rs
@@ -1,9 +1,9 @@
 pub const fn current_mainnet() -> &'static [u8] {
-    version_3_8_1()
+    version_3_9_0()
 }
 
 pub const fn current_testnet() -> &'static [u8] {
-    version_3_8_1()
+    version_3_9_0()
 }
 
 pub const fn version_2_2_0() -> &'static [u8; 566653] {
@@ -40,6 +40,10 @@ pub const fn version_3_7_0() -> &'static [u8; 1402916] {
 
 pub const fn version_3_8_1() -> &'static [u8; 1515608] {
     include_bytes!("../archive/signer-3_8_1.wasm")
+}
+
+pub const fn version_3_9_0() -> &'static [u8; 1481703] {
+    include_bytes!("../archive/signer-3_9_0.wasm")
 }
 
 #[cfg(test)]
